@@ -41,27 +41,27 @@
 			label="Table View"
 			disabled={viewType == 'table'}
 			variant="secondary"
-			on:click={() => (viewType = 'table')}
+			onclick={() => (viewType = 'table')}
 		>
-			<i slot="icon"><Table /></i>
+			{#snippet icon()}<Table />{/snippet}
 		</Button>
 
 		<Button
 			label="Graph View"
 			variant="secondary"
 			disabled={viewType == 'graph' || inFlight}
-			on:click={() => (viewType = 'graph')}
+			onclick={() => (viewType = 'graph')}
 		>
-			<i slot="icon"><Graph /></i>
+			{#snippet icon()}<Graph />{/snippet}
 		</Button>
 
 		<Button
 			label="TTL View"
 			variant="secondary"
 			disabled={viewType == 'ttl'}
-			on:click={() => (viewType = 'ttl')}
+			onclick={() => (viewType = 'ttl')}
 		>
-			<i slot="icon"><Turtle /></i>
+			{#snippet icon()}<Turtle />{/snippet}
 		</Button>
 	</div>
 
