@@ -45,10 +45,10 @@
 				placeholder="Select a local data source"
 				value={selectedSourceId}
 				on:icChange={handleSourceChange}
-			/>
+			></ic-select>
 
 			<div class="align-bottom inline">
-				<Button label="Persist" {disabled} on:click={handlePersist} />
+				<Button label="Persist" {disabled} onclick={handlePersist} />
 			</div>
 
 			{#if persisted}
