@@ -28,10 +28,10 @@
 		};
 	};
 
-	enum LocalDataSourceImportTab {
-		Document = 0,
-		PrefixBrowser = 1
-	}
+	const LocalDataSourceImportTab = {
+		Document: 0,
+		PrefixBrowser: 1
+	};
 
 	// State
 	let document = '';
@@ -93,7 +93,7 @@
 		on:icChange={handleImportTypeChanged}
 		label="Import format"
 		options={Object.entries(imports).map(([value, { label }]) => ({ label, value }))}
-	/>
+	></ic-select>
 
 	<TabNavigation bind:selectedTabIndex>
 		<Tab title="Document" />
